LayoutTests/imported/w3c/web-platform-tests/cssom/getComputedStyle-pseudo-expected.txt


PASS Resolution of width is correct for ::before and ::after pseudo-elements 
FAIL Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements assert_equals: expected "50px" but got "50%"
PASS Resolution of nonexistent pseudo-element styles 
FAIL Resolution of pseudo-element styles in display: none elements assert_equals: Pseudo-styles of display: none elements should be correct expected "\"Foo\"" but got "Foo"
